food routes ok

+++ Backend/orderservice/src/main/java/com/deliverysystem/orderservice/repository/OrderRepository.java
@@ -4,8 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
     // Additional query methods can be defined here if needed
-    
 }
